update Javascript files
