+++ index.html
@@ -5,8 +5,8 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     
-    <link rel="stylesheet" href="styles.css" />
-    <script src="app.js" defer></script>
+    <link rel="stylesheet" href="index.css" />
+    <script src="index.js" defer type="module"></script>
     <title>Awesome Books Website</title>
   </head>
   <header id="header">
@@ -52,6 +52,5 @@
       <p>Copyright ...</p>
     </footer>
     <script src="https://cdn.jsdelivr.net/npm/luxon@1.27.0/build/global/luxon.min.js" crossorigin="anonymous"></script>
-
   </body>
 </html>
